channel-members-tracker/pyproject.toml
[tool.poetry]
name = "channel-members-tracker"
version = "0.1.0"
description = ""
authors = ["Daniil <daniil@vivaldi.net>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
Telethon = "^1.25.4"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
pika = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tracker = "src.main:launch"
consumer = "src.main:consume"
generate-session-key = "src.generate_session_key:launch"

